backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
python-jose==3.3.0
bcrypt==4.0.1
openai==1.3.7
pillow==10.1
reportlab==4.0.7
python-pptx==1.0.2
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
torch==2.6.0